requirements.txt
a2wsgi==1.10.8
anyio==4.8.0
apscheduler==3.10.1
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
connexion==3.2.0
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kazoo==2.5.0
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pykafka==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
swagger_ui_bundle==1.1.0
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
